Now appyling trim-full-walk in the result of levy-foodwalk, straight-foodwalk
This removes redundant info in the third element of the result,
the full walk, that is already in the second element, the walk
until food was found.  I generally only use the remaining elements
of full-walk to show where the foodwalk would have gone after it
found food, so there's no reason to keep the overlap.  It's simple
enough to reconstruct the full walk again from the remaining data.