Revised did-couldve-foodwalk-plot to always plot the full walk.
Previously, the underlying full walk was not plotted if no food
was found.  Up until recently, in this case the full walk was
identity to the food search walk, since no food was found.
However, now levy-foodwalk is calling trim-full-walk.  So if
no food was found, the third term its return value is just
an empty sequence, so it does no harm to "plot it". Doing so
should preserve the labeling colors across plots with and without
found food.