Fixed bug in trim-full-walk.
Earlier I'd changed the overlap from 2 to 1.  That was a mistake.
Without an overlap of two between the walk that finds food and
its extension beyond the food, there's a gap between the location
from which food was found, and the beginning of the next segment.
I fixed this in this commit.