Fixed bug in find-in-seg, removed debugging code
See new comment above find-in-seg for details.  The problem was that
when a slope was very near 0.0, the y-eps increment could be so small
that adding it to y resulted in the same y, and as a result, the y shift
never got to the y endpoint of the segment.  Solution is to stop testing
for y = endpoint, since slopes are managed s.t. x is always changing
until the endpoint is reached.