Abstracted out coordinate-walk steps in levy-foodwalk in new function.
The new function is currently called 'foodwalk'.  It will be useful when
I want to pass a walk as a seqeuence of location coordinate pairs rather
than as a sequence of mathematical vectors representing steps.  This
will be useful e.g. when embedding Archimedean spirals in a walk (since
they are easiest to construct as sequences of locations).