Now initializing matrix cells with empty sets (vs. nils)
If I initialize env matrix cells by nil, and then conj data
onto the nil, the result is a PersistentList.  When this is
in the first column of the matrix, you can't use mx/pm to
display the matrix, because core.matrix's internal get-shape
function tries to recurse into the first column in order to
find out the dimensionality of the matrix.  It uses count
to do this.  count doesn't work on sets, so core.matrix isn't
confused.  See the issue I submitted:
https://github.com/mikera/core.matrix/issues/361