Renamed take-rand to take-rand!
"take-rand!" highlights that it modifies the buffer in
its argument.

The standard take function, when applied to
a lazy sequence of random numbers generated on the fly also
modifies the PRNG, so that's in effect imperative as well.  So for purposes
of this repo, take-rand is in a sense no more imperative than
the way I'm using take which was the reason for leaving the bang off
of the name.  But I think it's important to keep the bang, especially
as I further develop code arounbd take-rand!